pyproject.toml
[tool.poetry]
name = "monitor-cli"
version = "1.0.0"
description = "CLI used to control container monitoring service"
authors = ["Masatoshi Tanida <mtanida@outlook.com>"]
license = "NONE"
packages = [
  { include = "monitorcli" }
]

[tool.poetry.scripts]
monitor-cli = 'monitorcli.cli:app'

[tool.poetry.dependencies]
python = "^3.7"
typer = "^0.9.0"
requests-unixsocket = "^0.3.0"
urllib3 = "<2.0.0"
pyyaml = "^6.0.1"

[tool.poetry.dev-dependencies]
black = "22.3.0"
isort = "5.10.1"
pydocstyle = "^6.3.0"

[tool.black]
line-length = 100  # Set your preferred maximum line length here

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
